LayoutTests/js/dom/array-sort-non-writeable-proto-hole-expected.txt

PASS array.sort() threw exception TypeError: Attempted to assign to readonly property..
PASS successfullyParsed is true

TEST COMPLETE
